+++ packages/welcome/lib/changelog-view.js
@@ -60,7 +60,7 @@ export default class ChangeLogView {
                 </ul>
               </li>
               <li>
-                [settings-view] Fix issue that closes project windows when the user clicked on a badge in a package card.
+                [settings-view] Fix issue that closes project windows when the user clicks on a badge in a package card.
               </li>
               <li>
                 [find-and-replace] Fix issue that prevents searches with patterns from working properly in projects with multiple roots.
@@ -75,10 +75,10 @@ export default class ChangeLogView {
                 Fix macOS issue causing buffer contents to fail to update for some users when modified by another program.
               </li>
               <li>
-                Fix issue causing the <code>.deb</code> installation to refuse to launch because of improper permissions.
+                Fix issue causing the <code>.deb</code> installation to refuse to launch because of improper permissions/AppArmor profile.
               </li>
               <li>
-                Fix issue causing the <code>.rpm</code> installation to run the wrong script when upgrading.
+                Fix issue causing the <code>.rpm</code> installation to run the wrong script when upgrading (restore pulsar and ppm to PATH).
               </li>
               <li>
                 Fix issue causing the <code>--package</code> switch not to work correctly when invoked directly on the Pulsar binary, rather than on <code>pulsar.sh</code>/<code>pulsar.cmd</code>.